pyproject.toml
[project]
name = "bc-framework"
version = "0.1.0"
description = "A business-oriented Python framework designed for rapid development of microservices and workers, simplifying integrations and providing built-in solutions for security and data access."
keywords = [
    "microservice",
    "worker",
    "framework",
    "business-oriented",
    "api",
    "rest",
    "authentication",
    "authorization",
    "integration",
    "connector",
    "distributed-systems",
    "asynchronous",
    "scalable",
    "enterprise",
    "toolkit",
    "library",
    "fast-development",
    "data-access",
    "inter-service-communication",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Internet",
    "Topic :: System :: Distributed Computing",
    "Topic :: Utilities",
]

license = { text = "MIT" }
readme = "README.md"

authors = [
    { name = "Артём Широких", email = "job@artemetr.ru" },
    { name = "Esoft", email = "it@esoft.tech" },
]
maintainers = [{ name = "Артём Широких", email = "job@artemetr.ru" }]

requires-python = ">=3.12"
dependencies = ["bc-configs", "pymongo>=4.15.3"]

[project.urls]
Homepage = "https://github.com/etagi-esoft/py-bc-framework"
Source = "https://github.com/etagi-esoft/py-bc-framework"
Documentation = "https://esoft-tech.github.io/py-bc-framework/index.html"

[project.optional-dependencies]
hashicorp = ["bc-framework", "hvac"]

[dependency-groups]
docs = [
    "bc-framework",
    "autodoc-pydantic>=2.2.0",
    "furo>=2025.9.25",
    "myst-parser>=4.0.1",
    "sphinx>=8.2.3",
    "sphinx-book-theme>=1.1.4",
    "sphinx-rtd-theme>=3.0.2",
]
dev = [
    "bc-framework",
    "coverage",
    "mypy",
    "pre-commit",
    "python-dotenv",
    "ruff",
]
test = [
    "bc-framework",
    "pytest",
    "pytest-cov",
    "python-dotenv",
    "pytest-asyncio",
    "testcontainers[mongodb]",
]

[tool.uv]
default-groups = ["docs", "dev", "test"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bc_framework"]

[tool.hatch.build.targets.sdist]
include = ["src/bc_framework", "README.md", "LICENSE"]

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
show_missing = true
fail_under = 70
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

[tool.mypy]
plugins = ['pydantic.mypy']
exclude = ["^docs/*"]

[tool.pytest.ini_options]
pythonpath = [".", "src"]
